market_app/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "market-app"
version = "0.1.0"
description = "Offline-first market monitor and watchlist tooling."
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
  "pandas==2.2.3",
  "numpy==2.1.3",
  "pyarrow==17.0.0",
  "requests==2.32.4",
  "python-dateutil==2.9.0.post0",
  "PyYAML==6.0.2",
  "tomli==2.0.1; python_version < \"3.11\"",
]

[project.scripts]
market-monitor = "market_app.cli:main"

[project.optional-dependencies]
test = [
  "pytest==8.3.4",
  "scikit-learn==1.5.2",
]

dev = [
  "pandas==2.2.3",
  "numpy==2.1.3",
  "requests==2.32.4",
  "python-dateutil==2.9.0.post0",
  "PyYAML==6.0.2",
  "tomli==2.0.1",
  "scikit-learn==1.5.2",
  "matplotlib==3.9.2",
  "pytest==8.3.4",
  "ruff==0.7.3",
  "black==24.10.0",
  "mypy==1.12.1",
  "bandit==1.7.9",
  "pre-commit==3.8.0",
  "types-requests==2.32.0.20240914",
  "types-python-dateutil==2.9.0.20241003",
]

[tool.setuptools]
package-dir = {"" = ".", "market_app" = "src/market_app"}

[tool.setuptools.packages.find]
where = ["."]
include = ["market_monitor*", "market_app*"]

[tool.setuptools.package-data]
market_app = ["sample_data/**"]

[tool.black]
line-length = 100
target-version = ["py310"]
extend-exclude = '''
/(\\.venv|_backup|_backup_.*|data_cache|outputs|output)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
  ".venv",
  "_backup",
  "_backup_*",
  "data",
  "data_cache",
  "outputs",
  "output",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "SIM", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["market_monitor", "market_app"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
check_untyped_defs = false
disallow_incomplete_defs = false
exclude = ["_backup", "_backup_.*", "data", "data_cache", "outputs", "output"]

[tool.bandit]
skips = ["B101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
filterwarnings = [
  "ignore:'oneOf' deprecated - use 'one_of':DeprecationWarning",
  "ignore:'parseString' deprecated - use 'parse_string':DeprecationWarning",
  "ignore:'resetCache' deprecated - use 'reset_cache':DeprecationWarning",
  "ignore:'enablePackrat' deprecated - use 'enable_packrat':DeprecationWarning",
  "ignore:X has feature names, but SimpleImputer was fitted without feature names:UserWarning:sklearn",
]
